makeFile improvements

+++ Makefile
@@ -9,14 +9,29 @@ endif
 .PHONY: all browser
 
 all:
+	make deps
+	if [ ! -f owncloud/test/config.json ] ; then cp owncloud/test/config.sample.json owncloud/test/config.json ; fi;
+	if [ ! -f swagger.config.js ] ; then touch swagger.config.js ; fi;
+	echo "module.exports = {\n\towncloudURL: \"\",\n\tusername: \"\",\n\tpassword: \"\"\n};" > swagger.config.js
+	bash readOCInfo.sh
+
+deps:
 	sudo npm i
 	sudo npm --prefix ./docs-swagger/ i ./docs-swagger/
-	cp owncloud/test/config.sample.json owncloud/test/config.json
 
-swagger:
+depsSilent:
+	sudo npm i --silent
+	sudo npm --prefix ./docs-swagger/ i ./docs-swagger/ --silent
+
+swagger: depsSilent
+	if [ ! -f swagger.config.js ] ; then \
+		touch swagger.config.js ; \
+		echo "module.exports = {\n\towncloudURL: \"\",\n\tusername: \"\",\n\tpassword: \"\"\n};" > swagger.config.js ; \
+		bash readOCInfo.sh ; \
+	fi;
 	node docs-swagger/server.js
 
-test:
+test: depsSilent
 	if [ owncloud/test/testDownloadDir ] ; then rm -rf owncloud/test/testDownloadDir ; fi;
 	mkdir owncloud/test/testDownloadDir
 	if [ ! -f owncloud/test/config.json ] ; then cp owncloud/test/config.sample.json owncloud/test/config.json ; fi;
@@ -39,7 +54,7 @@ jsdocs:
 	#Output the final documentation link
 	echo "To read the documentation, click here : file://"${DIR}"/jsdoc/ownCloud.html"
 
-browser:
+browser: depsSilent
 	sed -i "s/require('fs/require('fs-web/g" owncloud/*.js
 	sed -i "s/require('fs/require('fs-web/g" owncloud/test/*.js
 
@@ -59,7 +74,8 @@ clean:
 	rm -rf jsdoc/
 	rm -rf node_modules/
 	rm -rf docs-swagger/node_modules/
-	rm owncloud/test/config.json
+	rm -rf owncloud/test/config.json
+	rm -rf swagger.config.js
 
 	#Output success message
-	echo "Existing Documentation removed"
+	echo "Repo cleaned, run \"make\" to setup again."
